requirements.txt
boto3
botocore
datetime
python-a2s
